// src/store/user/user.types.ts
export enum USER_ACTION_TYPES {
    SET_CURRENT_USER = 'user/SET_CURRENT_USER',
    CHECK_USER_SESSION = 'user/CHECK_USER_SESSION',
    GOOGLE_SIGNIN_START = 'user/GOOGLE_SIGNIN_START',
    EMAIL_SIGNIN_START = 'user/EMAIL_SIGNIN_START',
    SIGNIN_SUCCEESS = 'user/SIGNIN_SUCCEESS',
    SIGNIN_FAIL = 'user/SIGNIN_FAIL',

    SIGNUP_START = 'user/SIGNUP_START',
    SIGNUP_SUCCEESS = 'user/SIGNUP_SUCCEESS',
    SIGNUP_FAIL = 'user/SIGNUP_FAIL',

    SIGNOUT_START = 'user/SIGNOUT_START',
    SIGNOUT_SUCCESS = 'user/SIGNOUT_SUCCESS',
    SIGNOUT_FAIL = 'user/SIGNOUT_FAIL',

}